roster/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
Flask==3.0.3
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.6
psycopg2-binary==2.9.5
PyJWT==2.7.0
python-dotenv==0.21.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.4
Werkzeug==3.0.3
